+++ src/nationalarchives/components/breadcrumbs/breadcrumbs.stories.js
@@ -7,6 +7,7 @@ const argTypes = {
   items: { control: "object" },
   noCollapse: { control: "boolean" },
   structuredData: { control: "boolean" },
+  labelText: { control: "text" },
   classes: { control: "text" },
   attributes: { control: "object" },
 };
@@ -22,12 +23,20 @@ export default {
   argTypes,
 };
 
-const Template = ({ items, noCollapse, structuredData, classes, attributes }) =>
+const Template = ({
+  items,
+  noCollapse,
+  structuredData,
+  labelText,
+  classes,
+  attributes,
+}) =>
   Breadcrumbs({
     params: {
       items,
       noCollapse,
       structuredData,
+      labelText,
       classes,
       attributes,
     },